requirements.txt
fastapi>=0.115.0
uvicorn[standard]>0.30.0
boto3>=1.34.0
supabase>=2.4.0
python-dotenv>=1.0.1
pydantic-settings>=2.4.0